+++ test/src/Game/Werewolf/Test/Game.hs
@@ -33,13 +33,16 @@ allGameTests =
     , testProperty "new game starts on first round"                         prop_newGameStartsOnFirstRound
     , testProperty "new game starts with events empty"                      prop_newGameStartsWithEventsEmpty
     , testProperty "new game starts with passes empty"                      prop_newGameStartsWithPassesEmpty
-    , testProperty "new game starts with no heal"                           prop_newGameStartsWithNoHeal
-    , testProperty "new game starts with no heal used"                      prop_newGameStartsWithNoHealUsed
+    , testProperty "new game starts with allowed voters full"               prop_newGameStartsWithAllowedVotersFull
+    , testProperty "new game starts with heal false"                        prop_newGameStartsWithHealFalse
+    , testProperty "new game starts with heal used false"                   prop_newGameStartsWithHealUsedFalse
     , testProperty "new game starts with no poison"                         prop_newGameStartsWithNoPoison
-    , testProperty "new game starts with no poison used"                    prop_newGameStartsWithNoPoisonUsed
+    , testProperty "new game starts with poison used false"                 prop_newGameStartsWithPoisonUsedFalse
     , testProperty "new game starts with no prior protect"                  prop_newGameStartsWithNoPriorProtect
     , testProperty "new game starts with no protect"                        prop_newGameStartsWithNoProtect
+    , testProperty "new game starts with scapegoat blamed false"            prop_newGameStartsWithScapegoatBlamedFalse
     , testProperty "new game starts with no see"                            prop_newGameStartsWithNoSee
+    , testProperty "new game starts with village idiot revealed false"      prop_newGameStartsWithVillageIdiotRevealedFalse
     , testProperty "new game starts with votes empty"                       prop_newGameStartsWithVotesEmpty
     , testProperty "new game uses given players"                            prop_newGameUsesGivenPlayers
     ]
@@ -63,17 +66,20 @@ prop_newGameStartsWithEventsEmpty players = null $ newGame players ^. events
 prop_newGameStartsWithPassesEmpty :: [Player] -> Bool
 prop_newGameStartsWithPassesEmpty players = null $ newGame players ^. passes
 
-prop_newGameStartsWithNoHeal :: [Player] -> Bool
-prop_newGameStartsWithNoHeal players = not $ newGame players ^. heal
+prop_newGameStartsWithAllowedVotersFull :: [Player] -> Property
+prop_newGameStartsWithAllowedVotersFull players = newGame players ^. allowedVoters === map (view name) players
 
-prop_newGameStartsWithNoHealUsed :: [Player] -> Bool
-prop_newGameStartsWithNoHealUsed players = not $ newGame players ^. healUsed
+prop_newGameStartsWithHealFalse :: [Player] -> Bool
+prop_newGameStartsWithHealFalse players = not $ newGame players ^. heal
+
+prop_newGameStartsWithHealUsedFalse :: [Player] -> Bool
+prop_newGameStartsWithHealUsedFalse players = not $ newGame players ^. healUsed
 
 prop_newGameStartsWithNoPoison :: [Player] -> Bool
 prop_newGameStartsWithNoPoison players = isNothing $ newGame players ^. poison
 
-prop_newGameStartsWithNoPoisonUsed :: [Player] -> Bool
-prop_newGameStartsWithNoPoisonUsed players = not $ newGame players ^. poisonUsed
+prop_newGameStartsWithPoisonUsedFalse :: [Player] -> Bool
+prop_newGameStartsWithPoisonUsedFalse players = not $ newGame players ^. poisonUsed
 
 prop_newGameStartsWithNoPriorProtect :: [Player] -> Bool
 prop_newGameStartsWithNoPriorProtect players = isNothing $ newGame players ^. priorProtect
@@ -81,8 +87,15 @@ prop_newGameStartsWithNoPriorProtect players = isNothing $ newGame players ^. pr
 prop_newGameStartsWithNoProtect :: [Player] -> Bool
 prop_newGameStartsWithNoProtect players = isNothing $ newGame players ^. protect
 
+prop_newGameStartsWithScapegoatBlamedFalse :: [Player] -> Bool
+prop_newGameStartsWithScapegoatBlamedFalse players = not $ newGame players ^. scapegoatBlamed
+
 prop_newGameStartsWithNoSee :: [Player] -> Bool
 prop_newGameStartsWithNoSee players = isNothing $ newGame players ^. see
+
+prop_newGameStartsWithVillageIdiotRevealedFalse :: [Player] -> Bool
+prop_newGameStartsWithVillageIdiotRevealedFalse players =
+    not $ newGame players ^. villageIdiotRevealed
 
 prop_newGameStartsWithVotesEmpty :: [Player] -> Bool
 prop_newGameStartsWithVotesEmpty players = Map.null $ newGame players ^. votes